requirements.txt
# AWS Lambda requirements
boto3>=1.28.0
botocore>=1.31.0

# Email and messaging
email-validator>=2.0.0

# PDF generation (Lambda compatible)
reportlab>=4.0.0
pillow>=10.0.0

# HTTP requests
requests>=2.31.0

# Date/time handling
python-dateutil>=2.8.0

# JSON handling
simplejson>=3.19.0

# Email parsing
email-mime-parser>=1.0.0

# Google APIs (for Gmail)
google-api-python-client>=2.100.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# Utilities
python-dotenv>=1.0.0